--- packages/gamut/src/Tabs copy/Tab.tsx
@@ -1,32 +1,0 @@
-import styled from '@emotion/styled';
-import * as React from 'react';
-import {
-  Tab as ReactAriaTab,
-  TabProps as ReactAriaTabProps,
-} from 'react-aria-components';
-
-import { tabElementBaseProps } from './props';
-import { TabButtonProps, tabStates, tabVariants } from './TabButton';
-import { useTabShared } from './TabProvider';
-
-interface TabBaseProps extends TabButtonProps, ReactAriaTabProps {}
-
-export type TabProps = TabBaseProps &
-  Omit<ReactAriaTabProps, 'id'> & {
-    /**
-     * the id matches up the tab and tab panel
-     */
-    id: string;
-  };
-
-export const TabBase = styled(ReactAriaTab)<TabProps>(
-  tabVariants,
-  tabStates,
-  tabElementBaseProps
-);
-
-export const Tab: React.FC<TabProps> = (props) => {
-  const { variant } = useTabShared();
-
-  return <TabBase {...props} variant={variant} />;
-};
