--- packages/gamut/src/Tabs copy/TabPanels.tsx
@@ -1,7 +1,0 @@
-import styled from '@emotion/styled';
-
-import { tabElementBaseProps, TabElementStyleProps } from './props';
-
-export interface TabPanelsProps extends TabElementStyleProps {}
-
-export const TabPanels = styled('div')<TabPanelsProps>(tabElementBaseProps);
