--- packages/gamut/src/Tabs copy/index.tsx
@@ -1,8 +1,0 @@
-export * from './Tab';
-export * from './TabList';
-export * from './TabPanel';
-export * from './TabPanels';
-export * from './Tabs';
-export * from './TabNav';
-export * from './TabNavLink';
-export * from './TabButton';
