+++ packages/gamut/src/Tabs/TabPanel.tsx
@@ -1,19 +1,22 @@
-import { styledOptions } from '@codecademy/gamut-styles';
-import { StyleProps } from '@codecademy/variance';
 import styled from '@emotion/styled';
 import {
-  TabPanel as ReachTabPanel,
-  TabPanelProps as ReachTabPanelProps,
-} from '@reach/tabs';
+  TabPanel as ReactAriaTabPanel,
+  TabPanelProps as ReactAriaTabPanelProps,
+} from 'react-aria-components';
 
 import { tabElementBaseProps, TabElementStyleProps } from './props';
 
-export interface TabPanelProps
-  extends ReachTabPanelProps,
-    StyleProps<typeof tabElementBaseProps>,
-    TabElementStyleProps {}
+interface TabPanelBaseProps
+  extends TabElementStyleProps,
+    ReactAriaTabPanelProps {}
 
-export const TabPanel = styled(
-  ReachTabPanel,
-  styledOptions
-)(tabElementBaseProps);
+export type TabPanelProps = TabPanelBaseProps &
+  Omit<ReactAriaTabPanelProps, 'id'> & {
+    /**
+     * the id matches up the tab and tab panel
+     */
+    id: string;
+  };
+
+export const TabPanel =
+  styled(ReactAriaTabPanel)<TabPanelProps>(tabElementBaseProps);
