+++ packages/styleguide/src/lib/Molecules/Tabs/Tabs.mdx
@@ -38,7 +38,7 @@ By default, the tabs component is uncontrolled, meaning it will maintain it's ow
 
 ### Controlled
 
-When `selectedKey` is specified, the component switches to being a controlled component. It won't change its own tab state, but will instead notify of a request to change using `onSelectionChange`.
+Tabs can be controlled using the `selectedKey` prop, paired with the `onSelectionChange` prop. The `id` prop from the selected tab will be passed into the callback when the tab is selected, allowing you to update state accordingly.
 
 <Canvas of={TabStories.Controlled} />
 
@@ -46,7 +46,7 @@ When `selectedKey` is specified, the component switches to being a controlled co
 
 Tabs can use the Badge component to display a status, e.g. indicate new or updated content.
 
-<Canvas of={TabStories.WithBage} />
+<Canvas of={TabStories.WithBadge} />
 
 ### Block variant
